requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-login==0.6.2
flask-bcrypt==1.0.1
flask-sqlalchemy==3.0.3
gunicorn==20.1.0
opencv-python-headless==4.7.0.72
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
tensorflow-cpu==2.11.0


